prepInputsFireRegimePolys

Create fire regime polygons for scfmRegime.

Usage

fireRegimePolyTypes()

prepInputsFireRegimePolys(
  url = NULL,
  destinationPath = tempdir(),
  studyArea = NULL,
  rasterToMatch = NULL,
  type = "ECOREGION"
)

Arguments

url

character. URL from which to download and prepare fire regime polygons. Defaults are provided for Canadian ecodistrict, ecoregion, ecoprovince, and ecozone, as well as national Fire Regime Types and Fire Regime Units from Erni et al. (2020) doi:10.1139/cjfr-2019-0191 .

destinationPath

character. Path to directory where data will be downloaded.

studyArea

sf object corresponding to the study area of interest.

rasterToMatch

TODO

type

character. The polygon type to use: Must be one of "ECODISTRICT", "ECOREGION" (default), "ECOPROVINCE", "ECOZONE", "FRT", or "FRU". If url to BEC shapefile is provided, can also be one of: "BECNDT", "BECSUBZONE", or "BECZONE".

Examples

library(sf)
#> Linking to GEOS 3.10.2, GDAL 3.4.1, PROJ 8.2.1; sf_use_s2() is TRUE
library(sp)

## random study area in central Alberta
studyArea <- SpatialPoints(data.frame(lon = -115, lat = 55), proj4string = CRS("EPSG:4326")) %>%
st_as_sf() %>%
  st_transform(paste("+proj=lcc +lat_1=49 +lat_2=77 +lat_0=0 +lon_0=-95",
                   "+x_0=0 +y_0=0 +units=m +no_defs +ellps=GRS80 +towgs84=0,0,0")) %>%
  as_Spatial() %>%
  SpaDES.tools::randomStudyArea(center = ., seed = 60, size = 1e10) %>%
  st_as_sf()

# \donttest{
frpEcoregion <- prepInputsFireRegimePolys(studyArea = studyArea, type = "ECOREGION")
#> Running preProcess
#> Preparing: 
#> Checking local files...
#> Finished checking local files.
#> Error in packageVersion("googledrive") : 
#>   there is no package called ‘googledrive’
#> Error in packageVersion("googledrive") : 
#>   there is no package called ‘googledrive’
#> Error in packageVersion("googledrive") : 
#>   there is no package called ‘googledrive’
#> Error in downloadFile(archive = if (isTRUE(is.na(archive))) NULL else archive,     targetFile = targetFile, neededFiles = neededFiles, destinationPath = destinationPath,     quick = quick, checkSums = checkSums, dlFun = dlFun, url = url,     checksumFile = asPath(checkSumFilePath), needChecksums = needChecksums,     overwrite = overwrite, purge = purge, verbose = verbose,     .tempPath = .tempPath, ...): Download of  from https://sis.agr.gc.ca/cansis/nsdb/ecostrat/region/ecoregion_shp.zip failed. Please check the url that it is correct.
#> If the url is correct, it is possible that manually downloading it will work. To try this, with your browser, go to
#> https://sis.agr.gc.ca/cansis/nsdb/ecostrat/region/ecoregion_shp.zip,
#>  ... then download it manually, give it this name: 'ecoregion_shp.zip', and place file here: /tmp/RtmpsP00kR.
#> -------------------
#> If manual download was successful, you will likely also need to run Checksums manually after you download the file with this command: reproducible:::appendChecksumsTable(checkSumFilePath = '/tmp/RtmpsP00kR/CHECKSUMS.txt', filesToChecksum = '', destinationPath = '/tmp/RtmpsP00kR', append = TRUE)
plot(frpEcoregion)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'plot': object 'frpEcoregion' not found
# }

frpFRT <- prepInputsFireRegimePolys(studyArea = studyArea, type = "FRT")
#> Running preProcess
#> Preparing: 
#> Checking local files...
#> Finished checking local files.
#> Error in packageVersion("googledrive") : 
#>   there is no package called ‘googledrive’
#> Error in packageVersion("googledrive") : 
#>   there is no package called ‘googledrive’
#> Error in packageVersion("googledrive") : 
#>   there is no package called ‘googledrive’
#> Error in downloadFile(archive = if (isTRUE(is.na(archive))) NULL else archive,     targetFile = targetFile, neededFiles = neededFiles, destinationPath = destinationPath,     quick = quick, checkSums = checkSums, dlFun = dlFun, url = url,     checksumFile = asPath(checkSumFilePath), needChecksums = needChecksums,     overwrite = overwrite, purge = purge, verbose = verbose,     .tempPath = .tempPath, ...): Download of  from https://zenodo.org/record/4458156/files/FRT.zip failed. Please check the url that it is correct.
#> If the url is correct, it is possible that manually downloading it will work. To try this, with your browser, go to
#> https://zenodo.org/record/4458156/files/FRT.zip,
#>  ... then download it manually, give it this name: 'FRT.zip', and place file here: /tmp/RtmpsP00kR.
#> -------------------
#> If manual download was successful, you will likely also need to run Checksums manually after you download the file with this command: reproducible:::appendChecksumsTable(checkSumFilePath = '/tmp/RtmpsP00kR/CHECKSUMS.txt', filesToChecksum = '', destinationPath = '/tmp/RtmpsP00kR', append = TRUE)
plot(frpFRT)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'plot': object 'frpFRT' not found

frpFRU <- prepInputsFireRegimePolys(studyArea = studyArea, type = "FRU")
#> Running preProcess
#> Preparing: 
#> Checking local files...
#> Finished checking local files.
#> Error in packageVersion("googledrive") : 
#>   there is no package called ‘googledrive’
#> Error in packageVersion("googledrive") : 
#>   there is no package called ‘googledrive’
#> Error in packageVersion("googledrive") : 
#>   there is no package called ‘googledrive’
#> Error in downloadFile(archive = if (isTRUE(is.na(archive))) NULL else archive,     targetFile = targetFile, neededFiles = neededFiles, destinationPath = destinationPath,     quick = quick, checkSums = checkSums, dlFun = dlFun, url = url,     checksumFile = asPath(checkSumFilePath), needChecksums = needChecksums,     overwrite = overwrite, purge = purge, verbose = verbose,     .tempPath = .tempPath, ...): Download of  from https://zenodo.org/record/4458156/files/FRU.zip failed. Please check the url that it is correct.
#> If the url is correct, it is possible that manually downloading it will work. To try this, with your browser, go to
#> https://zenodo.org/record/4458156/files/FRU.zip,
#>  ... then download it manually, give it this name: 'FRU.zip', and place file here: /tmp/RtmpsP00kR.
#> -------------------
#> If manual download was successful, you will likely also need to run Checksums manually after you download the file with this command: reproducible:::appendChecksumsTable(checkSumFilePath = '/tmp/RtmpsP00kR/CHECKSUMS.txt', filesToChecksum = '', destinationPath = '/tmp/RtmpsP00kR', append = TRUE)
plot(frpFRU)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'plot': object 'frpFRU' not found