Various tools for a (right) Truncated Exponential Distribution

\(Z ~ TE(\theta, T)\): pdf is \(\frac{\theta \exp(-z \theta)}{1 - \exp(-T\theta)}\)

Usage

pTE(x, theta, T)

dTE(x, theta, T)

qTE(p, theta, T)

rTE(n, theta, T)

qqTE(
  x,
  theta = stop("missing theta"),
  T = stop("missing T"),
  plot.it = TRUE,
  xlab = deparse(substitute(x)),
  ylab = deparse(substitute(y)),
  ...
)

ETE(theta, T, a = 0)

ETEx(theta, T, a = 0, shift = 200)

ExpBar(Z)

HannonDayiha(Z, Tspec = 0)

Arguments

x

TODO

theta

TODO

T

TODO

p

TODO

n

TODO

plot.it

logical indicating whether to plot the resulting figure

xlab, ylab

label text for x- and y-axes

...

additional arguments passed to plot()

a

TODO

shift

TODO

Z

TODO

Tspec

TODO

Details

where T is the truncation point or upper bound and \(\theta\) is the shape parameter in this application, x are fire sizes >= shift, which is a lower bound and z = log(x / shift) are the scaled log transformed sizes which seem to fit a truncated exponential distribution fairly well.

Originally written by Steve in 1999 in support of Cumming CJFR 2001. Has been in use by BEACONs and was acquired from Pierre Vernier in May 17 2014.

pTE() is the distribution function;

dTE() is the density function;

qTE() is the quantile function;

rTE() is the random generation function. In fire size applications exp(rTE(n, theta, T)) * shift will generate n random fire sizes;

qqTE() produces a quantile-quantile plot of vector x against a TE(theta, T);

ETE() is TODO;

ETEx() is TODO;

ExpBar() is TODO;

HannonDayiha() implements the estimator of Hannon and Dayiha (1999), ported from 1999 C language implementation by SGC June 2004.

References

Patrick M. Hannona & Ram C. Dahiyaa (1999) Estimation of parameters for the truncated exponential distribution. Communications in Statistics - Theory and Methods 28(11): 2591-2612. doi:10.1080/03610929908832440 .